score functions into utils
